# 9130-contact-list/ContactList/ContactList/Entities/Person.cs
namespace ContactList.Entities;

public class Person
{
    public int Id { get; set; }
    public string? FirstName { get; set; }
    public string? LastName { get; set; }
    public string Email { get; set; } = string.Empty;
}